Rename persistence and model packages
