+++ routes/v0/contacts.py
@@ -1,7 +1,7 @@
 from flask import Blueprint, request, Response, abort, current_app
-from model import ApiError
-from model.contact import Contact
-from persistence.db import DBConnection
+from models.api import ApiError
+from models.contact import Contact
+from db.db import DBConnection
 from typing import Any, Optional, List
